requirements.txt
Flask==1.1.2
Flask-SocketIO==5.1.0
gevent==21.1.2
gevent-websocket==0.10.1
pymongo==3.11.4
python-engineio==4.2.0
python-socketio==5.3.0
Werkzeug==1.0.1
eventlet==0.31.0
aiohttp>=3.7.4
motor